bug fix, typo and rename
