+++ src/util/misc/autocomplete.ts
@@ -1,6 +1,6 @@
 import { AutocompleteInteraction } from "discord.js";
 import { DateTime, Duration } from "luxon";
-import { getTimezonefromRole } from "./role";
+import { getTimezoneFromRole } from "./role";
 
 export async function autocompleteISOTime(interaction: AutocompleteInteraction) {
 	const focus = interaction.options.getFocused().toUpperCase();
@@ -15,7 +15,7 @@ export async function autocompleteISOTime(interaction: AutocompleteInteraction) 
 			response.push({ name: "[Invalid] " + focus, value: focus });
 		}
 	} else {
-		const timezone = await getTimezonefromRole(interaction.user.id, interaction.guild!);
+		const timezone = await getTimezoneFromRole(interaction.user.id, interaction.guild!);
 		const now = DateTime.now()
 			.set({ hour: DateTime.now().hour + 1, minute: 0, second: 0, millisecond: 0 })
 			.setZone(timezone || "UTC")
